Fix pull request remarks - removed doctrine/migration dependency - fixed template
