+++ src/Provider/SchemaProviderInterface.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Okvpn\Bundle\MigrationBundle\Provider;
+
+interface SchemaProviderInterface
+{
+    /**
+     * Create the schema to which the database should be migrated.
+     *
+     * @return  \Doctrine\DBAL\Schema\Schema
+     */
+    public function createSchema();
+}
